♻️ Move startMonitoringBatch in its own file
